Fix more 2.13 warnings
Moved 'full application' examples to their own files
