+++ docs/src/test/scala/docs/http/scaladsl/HttpClientCollectingHeaders.scala
@@ -1,0 +1,26 @@
+package docs.http.scaladsl
+
+import akka.actor.ActorSystem
+import akka.http.scaladsl.Http
+import akka.http.scaladsl.model.headers.`Set-Cookie`
+import akka.http.scaladsl.model._
+
+import scala.concurrent.ExecutionContextExecutor
+import scala.concurrent.Future
+
+class HttpClientCollectingHeaders {
+  def main(args: Array[String]): Unit = {
+    implicit val system: ActorSystem = ActorSystem()
+    implicit val executionContext: ExecutionContextExecutor = system.dispatcher
+
+    val responseFuture: Future[HttpResponse] = Http().singleRequest(HttpRequest(uri = "http://akka.io"))
+
+    responseFuture.map {
+      case response @ HttpResponse(StatusCodes.OK, _, _, _) =>
+        val setCookies = response.headers[`Set-Cookie`]
+        println(s"Cookies set by a server: $setCookies")
+        response.discardEntityBytes()
+      case _ => sys.error("something wrong")
+    }
+  }
+}
