+++ docs/src/test/scala/docs/http/scaladsl/HttpClientSingleRequest.scala
@@ -1,0 +1,24 @@
+package docs.http.scaladsl
+
+import akka.actor.ActorSystem
+import akka.http.scaladsl.Http
+import akka.http.scaladsl.model._
+
+import scala.concurrent.Future
+import scala.util.{ Failure, Success }
+
+object HttpClientSingleRequest {
+  def main(args: Array[String]): Unit = {
+    implicit val system = ActorSystem()
+    // needed for the future flatMap/onComplete in the end
+    implicit val executionContext = system.dispatcher
+
+    val responseFuture: Future[HttpResponse] = Http().singleRequest(HttpRequest(uri = "http://akka.io"))
+
+    responseFuture
+      .onComplete {
+        case Success(res) => println(res)
+        case Failure(_)   => sys.error("something wrong")
+      }
+  }
+}
