+++ docs/src/test/scala/docs/http/scaladsl/HttpServerExampleSpec.scala
@@ -42,38 +42,6 @@ class HttpServerExampleSpec extends AnyWordSpec with Matchers
     //#binding-example
   }
 
-  "binding-failure-high-level-example" in compileOnlySpec {
-    //#binding-failure-high-level-example
-    import akka.actor.ActorSystem
-    import akka.http.scaladsl.Http
-    import akka.http.scaladsl.Http.ServerBinding
-    import akka.http.scaladsl.server.Directives._
-
-    import scala.concurrent.Future
-
-    object WebServer {
-      def main(args: Array[String]): Unit = {
-        implicit val system = ActorSystem()
-        // needed for the future foreach in the end
-        implicit val executionContext = system.dispatcher
-
-        val handler = get {
-          complete("Hello world!")
-        }
-
-        // let's say the OS won't allow us to bind to 80.
-        val (host, port) = ("localhost", 80)
-        val bindingFuture: Future[ServerBinding] =
-          Http().bindAndHandle(handler, host, port)
-
-        bindingFuture.failed.foreach { ex =>
-          log.error(ex, "Failed to bind to {}:{}!", host, port)
-        }
-      }
-    }
-    //#binding-failure-high-level-example
-  }
-
   // mock values:
   val handleConnections = {
     import akka.stream.scaladsl.Sink
@@ -213,125 +181,6 @@ class HttpServerExampleSpec extends AnyWordSpec with Matchers
     //#full-server-example
   }
 
-  "low-level-server-example" in compileOnlySpec {
-    //#low-level-server-example
-    import akka.actor.ActorSystem
-    import akka.http.scaladsl.Http
-    import akka.http.scaladsl.model.HttpMethods._
-    import akka.http.scaladsl.model._
-    import scala.io.StdIn
-
-    object WebServer {
-
-      def main(args: Array[String]): Unit = {
-        implicit val system = ActorSystem()
-        // needed for the future map/flatmap in the end
-        implicit val executionContext = system.dispatcher
-
-        val requestHandler: HttpRequest => HttpResponse = {
-          case HttpRequest(GET, Uri.Path("/"), _, _, _) =>
-            HttpResponse(entity = HttpEntity(
-              ContentTypes.`text/html(UTF-8)`,
-              "<html><body>Hello world!</body></html>"))
-
-          case HttpRequest(GET, Uri.Path("/ping"), _, _, _) =>
-            HttpResponse(entity = "PONG!")
-
-          case HttpRequest(GET, Uri.Path("/crash"), _, _, _) =>
-            sys.error("BOOM!")
-
-          case r: HttpRequest =>
-            r.discardEntityBytes() // important to drain incoming HTTP Entity stream
-            HttpResponse(404, entity = "Unknown resource!")
-        }
-
-        val bindingFuture = Http().bindAndHandleSync(requestHandler, "localhost", 8080)
-        println(s"Server online at http://localhost:8080/\nPress RETURN to stop...")
-        StdIn.readLine() // let it run until user presses return
-        bindingFuture
-          .flatMap(_.unbind()) // trigger unbinding from the port
-          .onComplete(_ => system.terminate()) // and shutdown when done
-
-      }
-    }
-    //#low-level-server-example
-  }
-
-  "high-level-server-example" in compileOnlySpec {
-    //#high-level-server-example
-    import akka.actor.ActorSystem
-    import akka.http.scaladsl.Http
-    import akka.http.scaladsl.model.{ ContentTypes, HttpEntity }
-    import akka.http.scaladsl.server.Directives._
-    import scala.io.StdIn
-
-    object WebServer {
-      def main(args: Array[String]): Unit = {
-        implicit val system = ActorSystem()
-        // needed for the future flatMap/onComplete in the end
-        implicit val executionContext = system.dispatcher
-
-        val route =
-          get {
-            concat(
-              pathSingleSlash {
-                complete(HttpEntity(ContentTypes.`text/html(UTF-8)`, "<html><body>Hello world!</body></html>"))
-              },
-              path("ping") {
-                complete("PONG!")
-              },
-              path("crash") {
-                sys.error("BOOM!")
-              }
-            )
-          }
-
-        // `route` will be implicitly converted to `Flow` using `RouteResult.route2HandlerFlow`
-        val bindingFuture = Http().bindAndHandle(route, "localhost", 8080)
-        println(s"Server online at http://localhost:8080/\nPress RETURN to stop...")
-        StdIn.readLine() // let it run until user presses return
-        bindingFuture
-          .flatMap(_.unbind()) // trigger unbinding from the port
-          .onComplete(_ => system.terminate()) // and shutdown when done
-      }
-    }
-    //#high-level-server-example
-  }
-
-  "minimal-routing-example" in compileOnlySpec {
-    //#minimal-routing-example
-    import akka.actor.ActorSystem
-    import akka.http.scaladsl.Http
-    import akka.http.scaladsl.model._
-    import akka.http.scaladsl.server.Directives._
-    import scala.io.StdIn
-
-    object WebServer {
-      def main(args: Array[String]): Unit = {
-
-        implicit val system = ActorSystem("my-system")
-        // needed for the future flatMap/onComplete in the end
-        implicit val executionContext = system.dispatcher
-
-        val route =
-          path("hello") {
-            get {
-              complete(HttpEntity(ContentTypes.`text/html(UTF-8)`, "<h1>Say hello to akka-http</h1>"))
-            }
-          }
-
-        val bindingFuture = Http().bindAndHandle(route, "localhost", 8080)
-
-        println(s"Server online at http://localhost:8080/\nPress RETURN to stop...")
-        StdIn.readLine() // let it run until user presses return
-        bindingFuture
-          .flatMap(_.unbind()) // trigger unbinding from the port
-          .onComplete(_ => system.terminate()) // and shutdown when done
-      }
-    }
-    //#minimal-routing-example
-  }
-
   "long-routing-example" in compileOnlySpec {
     //#long-routing-example
     import akka.actor.{ ActorRef, ActorSystem }
@@ -441,128 +290,6 @@ class HttpServerExampleSpec extends AnyWordSpec with Matchers
       }
     )
     //#long-routing-example
-  }
-
-  "stream random numbers" in compileOnlySpec {
-    //#stream-random-numbers
-    import akka.actor.ActorSystem
-    import akka.stream.scaladsl._
-    import akka.util.ByteString
-    import akka.http.scaladsl.Http
-    import akka.http.scaladsl.model.{ HttpEntity, ContentTypes }
-    import akka.http.scaladsl.server.Directives._
-    import scala.util.Random
-    import scala.io.StdIn
-
-    object WebServer {
-
-      def main(args: Array[String]): Unit = {
-
-        implicit val system = ActorSystem()
-        // needed for the future flatMap/onComplete in the end
-        implicit val executionContext = system.dispatcher
-
-        // streams are re-usable so we can define it here
-        // and use it for every request
-        val numbers = Source.fromIterator(() =>
-          Iterator.continually(Random.nextInt()))
-
-        val route =
-          path("random") {
-            get {
-              complete(
-                HttpEntity(
-                  ContentTypes.`text/plain(UTF-8)`,
-                  // transform each number to a chunk of bytes
-                  numbers.map(n => ByteString(s"$n\n"))
-                )
-              )
-            }
-          }
-
-        val bindingFuture = Http().bindAndHandle(route, "localhost", 8080)
-        println(s"Server online at http://localhost:8080/\nPress RETURN to stop...")
-        StdIn.readLine() // let it run until user presses return
-        bindingFuture
-          .flatMap(_.unbind()) // trigger unbinding from the port
-          .onComplete(_ => system.terminate()) // and shutdown when done
-      }
-    }
-    //#stream-random-numbers
-  }
-
-  "interact with an actor" in compileOnlySpec {
-    //#actor-interaction
-    import akka.actor.{ Actor, ActorSystem, Props, ActorLogging }
-    import akka.http.scaladsl.Http
-    import akka.http.scaladsl.model.StatusCodes
-    import akka.http.scaladsl.server.Directives._
-    import akka.http.scaladsl.marshallers.sprayjson.SprayJsonSupport._
-    import akka.pattern.ask
-    import akka.util.Timeout
-    import spray.json.DefaultJsonProtocol._
-    import scala.concurrent.Future
-    import scala.concurrent.duration._
-    import scala.io.StdIn
-
-    object WebServer {
-
-      case class Bid(userId: String, offer: Int)
-      case object GetBids
-      case class Bids(bids: List[Bid])
-
-      class Auction extends Actor with ActorLogging {
-        var bids = List.empty[Bid]
-        def receive = {
-          case bid @ Bid(userId, offer) =>
-            bids = bids :+ bid
-            log.info(s"Bid complete: $userId, $offer")
-          case GetBids => sender() ! Bids(bids)
-          case _       => log.info("Invalid message")
-        }
-      }
-
-      // these are from spray-json
-      implicit val bidFormat = jsonFormat2(Bid)
-      implicit val bidsFormat = jsonFormat1(Bids)
-
-      def main(args: Array[String]): Unit = {
-        implicit val system = ActorSystem()
-        // needed for the future flatMap/onComplete in the end
-        implicit val executionContext = system.dispatcher
-
-        val auction = system.actorOf(Props[Auction], "auction")
-
-        val route =
-          path("auction") {
-            concat(
-              put {
-                parameter("bid".as[Int], "user") { (bid, user) =>
-                  // place a bid, fire-and-forget
-                  auction ! Bid(user, bid)
-                  complete((StatusCodes.Accepted, "bid placed"))
-                }
-              },
-              get {
-                implicit val timeout: Timeout = 5.seconds
-
-                // query the actor for the current auction state
-                val bids: Future[Bids] = (auction ? GetBids).mapTo[Bids]
-                complete(bids)
-              }
-            )
-          }
-
-        val bindingFuture = Http().bindAndHandle(route, "localhost", 8080)
-        println(s"Server online at http://localhost:8080/\nPress RETURN to stop...")
-        StdIn.readLine() // let it run until user presses return
-        bindingFuture
-          .flatMap(_.unbind()) // trigger unbinding from the port
-          .onComplete(_ => system.terminate()) // and shutdown when done
-
-      }
-    }
-    //#actor-interaction
   }
 
   "consume entity using entity directive" in compileOnlySpec {
