+++ docs/src/test/scala/docs/http/scaladsl/HttpServerHighLevel.scala
@@ -1,0 +1,38 @@
+package docs.http.scaladsl
+
+import akka.actor.ActorSystem
+import akka.http.scaladsl.Http
+import akka.http.scaladsl.model.{ ContentTypes, HttpEntity }
+import akka.http.scaladsl.server.Directives._
+import scala.io.StdIn
+
+object HttpServerHighLevel {
+  def main(args: Array[String]): Unit = {
+    implicit val system = ActorSystem()
+    // needed for the future flatMap/onComplete in the end
+    implicit val executionContext = system.dispatcher
+
+    val route =
+      get {
+        concat(
+          pathSingleSlash {
+            complete(HttpEntity(ContentTypes.`text/html(UTF-8)`, "<html><body>Hello world!</body></html>"))
+          },
+          path("ping") {
+            complete("PONG!")
+          },
+          path("crash") {
+            sys.error("BOOM!")
+          }
+        )
+      }
+
+    // `route` will be implicitly converted to `Flow` using `RouteResult.route2HandlerFlow`
+    val bindingFuture = Http().bindAndHandle(route, "localhost", 8080)
+    println(s"Server online at http://localhost:8080/\nPress RETURN to stop...")
+    StdIn.readLine() // let it run until user presses return
+    bindingFuture
+      .flatMap(_.unbind()) // trigger unbinding from the port
+      .onComplete(_ => system.terminate()) // and shutdown when done
+  }
+}
