+++ docs/src/test/scala/docs/http/scaladsl/SprayJsonExample2.scala
@@ -1,0 +1,82 @@
+package docs.http.scaladsl
+
+import akka.actor.ActorSystem
+import akka.http.scaladsl.Http
+import akka.Done
+import akka.http.scaladsl.server.Route
+import akka.http.scaladsl.server.Directives._
+import akka.http.scaladsl.model.StatusCodes
+// for JSON serialization/deserialization following dependency is required:
+// "com.typesafe.akka" %% "akka-http-spray-json" % "10.1.7"
+import akka.http.scaladsl.marshallers.sprayjson.SprayJsonSupport._
+import spray.json.DefaultJsonProtocol._
+
+import scala.io.StdIn
+
+import scala.concurrent.Future
+
+object SprayJsonExample2 {
+
+  // needed to run the route
+  implicit val system = ActorSystem()
+  // needed for the future map/flatmap in the end and future in fetchItem and saveOrder
+  implicit val executionContext = system.dispatcher
+
+  var orders: List[Item] = Nil
+
+  // domain model
+  final case class Item(name: String, id: Long)
+  final case class Order(items: List[Item])
+
+  // formats for unmarshalling and marshalling
+  implicit val itemFormat = jsonFormat2(Item)
+  implicit val orderFormat = jsonFormat1(Order)
+
+  // (fake) async database query api
+  def fetchItem(itemId: Long): Future[Option[Item]] = Future {
+    orders.find(o => o.id == itemId)
+  }
+  def saveOrder(order: Order): Future[Done] = {
+    orders = order match {
+      case Order(items) => items ::: orders
+      case _            => orders
+    }
+    Future { Done }
+  }
+
+  def main(args: Array[String]): Unit = {
+
+    val route: Route =
+      concat(
+        get {
+          pathPrefix("item" / LongNumber) { id =>
+            // there might be no item for a given id
+            val maybeItem: Future[Option[Item]] = fetchItem(id)
+
+            onSuccess(maybeItem) {
+              case Some(item) => complete(item)
+              case None       => complete(StatusCodes.NotFound)
+            }
+          }
+        },
+        post {
+          path("create-order") {
+            entity(as[Order]) { order =>
+              val saved: Future[Done] = saveOrder(order)
+              onComplete(saved) { done =>
+                complete("order created")
+              }
+            }
+          }
+        }
+      )
+
+    val bindingFuture = Http().bindAndHandle(route, "localhost", 8080)
+    println(s"Server online at http://localhost:8080/\nPress RETURN to stop...")
+    StdIn.readLine() // let it run until user presses return
+    bindingFuture
+      .flatMap(_.unbind()) // trigger unbinding from the port
+      .onComplete(_ => system.terminate()) // and shutdown when done
+
+  }
+}
